backend/requirements.txt
google-cloud-firestore
python-dotenv
google-cloud-storage
requests
temporalio
Pillow
transformers
torch
opencv-python-headless
timm
psutil
pytest
scipy>=1.11.0
matplotlib>=3.7.0
pandas>=2.0.0
seaborn>=0.12.0